+++ src/apps/obsidian-plugin/views/RightSidebarReactView/AccountingSection/AccountsListItem.tsx
@@ -1,5 +1,4 @@
 import {
-	DateValueObject,
 	PriceValueObject,
 	StringValueObject,
 } from "@juandardilag/value-objects";
@@ -21,12 +20,10 @@ import {
 	SettingsIcon,
 	Trash2,
 } from "lucide-react";
-import { useContext, useEffect, useState } from "react";
-import { Currency } from "../../../../../contexts/Currencies/domain/currency.vo";
+import { useContext, useState } from "react";
 import { Input } from "../../../components/Input/Input";
 import { PriceInput } from "../../../components/Input/PriceInput";
 import { AccountsContext, AppContext, TransactionsContext } from "../Contexts";
-import { ExchangeRatesContext } from "../Contexts/ExchangeRatesContext";
 
 export const AccountsListItem = ({ account }: { account: Account }) => {
 	const { plugin } = useContext(AppContext);
@@ -38,9 +35,6 @@ export const AccountsListItem = ({ account }: { account: Account }) => {
 	const {
 		useCases: { adjustAccount },
 	} = useContext(TransactionsContext);
-	const {
-		useCases: { getExchangeRate },
-	} = useContext(ExchangeRatesContext);
 
 	const [adjustingBalance, setAdjustingBalance] = useState(false);
 	const [newBalance, setNewBalance] = useState(
@@ -93,45 +87,6 @@ export const AccountsListItem = ({ account }: { account: Account }) => {
 		}
 	};
 
-	const [convertedBalance, setConvertedBalance] =
-		useState<PriceValueObject | null>(null);
-
-	useEffect(() => {
-		const fetchConvertedBalance = async () => {
-			console.log("Fetching converted balance for account", {
-				accountId: account.id,
-				accountCurrency: account.currency.toString(),
-				defaultCurrency: plugin.settings.defaultCurrency,
-			});
-			if (account.currency.value === plugin.settings.defaultCurrency) {
-				return null;
-			}
-
-			const exchangeRate = await getExchangeRate.execute({
-				fromCurrency: account.currency,
-				toCurrency: new Currency(plugin.settings.defaultCurrency),
-				date: DateValueObject.createNowDate(),
-			});
-
-			if (!exchangeRate) {
-				return null;
-			}
-
-			const convertedValue =
-				exchangeRate.rate.value * account.balance.value.value;
-
-			setConvertedBalance(
-				new PriceValueObject(convertedValue, {
-					decimals: 2,
-					withSign: true,
-					withZeros: true,
-				})
-			);
-		};
-
-		fetchConvertedBalance();
-	}, [account, getExchangeRate, plugin.settings.defaultCurrency]);
-
 	return (
 		<>
 			<Accordion style={{ width: "100%" }}>
@@ -154,8 +109,8 @@ export const AccountsListItem = ({ account }: { account: Account }) => {
 						plugin.settings.defaultCurrency
 							? ""
 							: account.currency.value}{" "}
-						{convertedBalance
-							? `(≈ ${convertedBalance.toString()} ${
+						{account.defaultCurrencyBalance
+							? `(≈ ${account.defaultCurrencyBalance.value.toString()} ${
 									plugin.settings.defaultCurrency
 							  })`
 							: ""}