+++ src/contexts/Reports/domain/accounts-report.entity.ts
@@ -8,7 +8,10 @@ export class AccountsReport {
 		return this._accounts
 			.filter((acc) => acc.type.isAsset())
 			.reduce(
-				(total, acc) => total.plus(acc.balance.value),
+				(total, acc) =>
+					total.plus(
+						acc.defaultCurrencyBalance?.value ?? acc.balance.value
+					),
 				ReportBalance.zero()
 			);
 	}
@@ -17,7 +20,10 @@ export class AccountsReport {
 		return this._accounts
 			.filter((acc) => acc.type.isLiability())
 			.reduce(
-				(total, acc) => total.plus(acc.balance.value),
+				(total, acc) =>
+					total.plus(
+						acc.defaultCurrencyBalance?.value ?? acc.balance.value
+					),
 				ReportBalance.zero()
 			);
 	}